# api/swagger.yaml
swagger: "2.0"
info:
  description: ""
  version: "1.0.0"
  title: "WSS - API"
  contact:
    email: "citron0137@gmail.com"
host: "dochoon.com"
basePath: "/api/"
schemes:
- "http"
paths:
...
  /file:
    post:
      security:
      - UserSecurity: []         
      tags:
      - "File"
      summary: "Create File"
      description: ""
      operationId: "create File"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "file"
        required: true
        schema:
          type: string
      responses:
        200:
          description: "success operation"
          schema:
            $ref: "#/definitions/file"
        400:
          description: "failed operation"
          schema:
            $ref: "#/definitions/onError"
    get:
      security:
      - UserSecurity: []         
      tags:
      - "File"
      summary: "Get Files"
      description: ""
      operationId: "getFiles"
      produces:
      - "application/json"
      responses:
        200:
          description: "success operation"
          schema:
            type: array
            items:
              $ref: "#/definitions/file"
        400:
          description: "failed operation"
          schema:
            $ref: "#/definitions/onError"
  /file/{name}:
    get:
      security:
      - UserSecurity: []         
      tags:
      - "File"
      summary: "Get File"
      description: ""
      operationId: "getFile"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: name
        type: string
        required: true
      responses:
        200:
          description: "success operation"
          schema:
            $ref: "#/definitions/file"
        400:
          description: "failed operation"
          schema:
            $ref: "#/definitions/onError"
    put:
      deprecated: true
      security:
      - UserSecurity: []         
      tags:
      - "File"
      summary: "Get File"
      description: ""
      operationId: "update File"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: name
        type: string
        required: true
      - in: "body"
        name: "file"
        required: true
        schema:
          type: string
      responses:
        200:
          description: "success operation"
          schema:
            $ref: "#/definitions/file"
        400:
          description: "failed operation"
          schema:
            $ref: "#/definitions/onError"
    delete:
      security:
      - UserSecurity: []         
      tags:
      - "File"
      summary: "delete File"
      description: ""
      operationId: "delete File"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: name
        type: string
        required: true
      responses:
        200:
          description: "success operation"
          schema:
            $ref: "#/definitions/onError"
        400:
          description: "failed operation"
          schema:
            $ref: "#/definitions/onError"
...
definitions : 
  onError:
    properties:
      message: 
        type: string
      boolean:
        type: boolean
...
  file:
    properties:
      ix:
        type: number
      type:
        type: string
      user_ix:
        type: string
      is_private:
        type: boolean
      is_anon:
        type: boolean
      name:
        type: string
      create_at:
        type: string
      update_at:
        type: string
...
securityDefinitions:
  UserSecurity:
    type: apiKey
    in: header
    name: x-access-token
    schemes: -http
security:
  - UserSecurity: []         
...